Add reCAPTCHA spam protection to the contact form
Hi guys,

as I set up the Getform form for the contact page, I put my mail to receive notifications about new messages. Whenever there is some message, I forward it to the other team members, so you should have an impression about the number of messages the contact form generates (if I had to guess, maybe one message every month or two?). 

However, aside from useful messages, which I forward, there is a large number of spam messages, which all need to be checked in order to not miss the few useful messages. I'd estimate the number of spam messages to around 4-5 per day lately, so it is quite time consuming to deal with them. 

I'd propose to either embed some kind of spam protection or to delete the contact form completely, as we have our issues tracked here on Github. Here's an attempt to set up reCAPTCHA as a spam protection mechanism.